Implement custom node visitors with spec collaborator and custom matcher resolver
* Some experiments with anaon classes

* Implement custom parser with spec collaborator and custom matcher resolver

+++ spec/PhpSpec/Event/SpecificationEventSpec.php
@@ -56,6 +56,6 @@ class SpecificationEventSpec extends ObjectBehavior
     {
         $this->beConstructedWith($specification);
 
-        $this->getTime()->shouldReturn((double) 0.0);
+        $this->getTime()->shouldReturn(0.0);
     }
 }
